resizer logic refactored
