+++ src/Main.jsx
@@ -42,6 +42,7 @@ import usePreventScrolling from './hooks/usePreventScrolling';
 
 export const Main = () => {
   const index = useRecoilValue(mainIndex);
+
   const [isMounted, setIsMounted] = useState(false);
   // const [index, setIndex] = useState(mainIndex);
   const location = useLocation();
