+++ src/components/hits/components/HitsSkeletonLoader.jsx
@@ -1,8 +1,9 @@
-import useScreenSize from '@/hooks/useScreenSize';
 import CustomSkeleton from '@/components/skeletons/CustomSkeleton';
+import { windowSize } from '@/hooks/useScreenSize';
+import { useRecoilValue } from 'recoil';
 
 const SkeletonLoader = ({type}) => {
-  const { tablet, mobile } = useScreenSize();
+  const { tablet, mobile } = useRecoilValue(windowSize);
 
   // Change this number to render more placeholders on the SRP loader
   const resultsNumber = 20;
